[update] socket 코드
