+++ controller/chatRooms.js
@@ -11,18 +11,19 @@ async function chatRooms(req, res) {
         const existPost = await recruitPost.findOne({postId: Number(postId), nickname: nickname}); // 게시글-닉네임 존재 여부 확인위함
         const existPostId = await recruitPost.findOne({postId: Number(postId)}); // 게시글 번호 존재여부 확인 위함
         const existRoom = await chatRoom.findOne({postId: Number(postId), nickname: nickname}); // 방 존재 여부 확인위함
+
         console.log(postId);
         console.log(existPost);
         console.log(existPostId);
         console.log(existRoom);
 
         // 이미 채팅방 만들어져있는 경우
-        if ( existRoom ) {
-           return res.status(400).send({
-                result: "false",
-                message: "이미 만들어진 채팅방이 존재합니다."
-            });
-        }
+        // if ( existRoom ) {
+        //    return res.status(400).send({
+        //         result: "false",
+        //         message: "이미 만들어진 채팅방이 존재합니다."
+        //     });
+        // }
 
         // 본인이 본인 채팅방 들어가는 경우
         if ( existPost ) {
@@ -45,12 +46,13 @@ async function chatRooms(req, res) {
                 nickname,
                 postNickname: existPostId.nickname,
                 postTitle: existPostId.title
-        });
+        })
             console.log(createdChats);
             
         return res.status(200).send({
                 result: "true",
-                message: "채팅방이 생성되었습니다."
+                message: "채팅방이 생성되었습니다.",
+                roomId: createdChats.roomId
         });
     } catch (err) {
         return res.status(400).send({
