+++ package.json
@@ -16,6 +16,7 @@
     "nodemon": "^2.0.18",
     "passport": "^0.6.0",
     "passport-kakao": "^1.0.1",
-    "passport-local": "^1.0.0"
+    "passport-local": "^1.0.0",
+    "socket.io": "^4.5.1"
   }
 }
